Use the native address of render process objects as remote proxy identifier.

+++ ChromiumFX/Source/Remote/CfrBase.cs
@@ -37,7 +37,7 @@ namespace Chromium.Remote {
     /// Base class for all remote wrapper classes for ref counted CEF structs.
     /// </summary>
     public abstract class CfrBase : CfrObject {
-        internal CfrBase(ulong proxyId) : base(proxyId) { }
+        internal CfrBase(IntPtr proxyId) : base(proxyId) { }
     }
 
 }
