+++ ChromiumFX/Source/Remote/CfrStructure.cs
@@ -37,6 +37,6 @@ namespace Chromium.Remote {
     /// Base class for all remote wrapper classes for CEF structs without refcount.
     /// </summary>
     public abstract class CfrStructure : CfrObject {
-        internal CfrStructure(ulong proxyId) : base(proxyId) { }
+        internal CfrStructure(IntPtr proxyId) : base(proxyId) { }
     }
 }
